Default magnification level in Gnome Magnifier
I propose to change the default magnification level in GNOME Magnifier from 2.0 to 1.0. In the tool, changing the magnification level to a value different from 1.0 would turn on the Magnifier. (GNOME's Magnifier seems to have a default of "2.0", but it also has a separate button to turn the Magnifier on, while the N&P Gathering form only has the magnification slider to turn magnification on or off.)